feat: adding new unified endpoint to fetch resources
