+++ dashboard/components/inventory/components/InventorySidePanel.tsx
@@ -3,7 +3,11 @@ import providers from '../../../utils/providerHelper';
 import Button from '../../button/Button';
 import Sidepanel from '../../sidepanel/Sidepanel';
 import SidepanelTabs from '../../sidepanel/SidepanelTabs';
-import { InventoryItem, Pages, Tag } from '../hooks/useInventory';
+import {
+  InventoryItem,
+  Pages,
+  Tag
+} from '../hooks/useInventory/types/useInventoryTypes';
 import InventoryTagWrapper from './InventoryTagWrapper';
 
 type InventorySidePanelProps = {
