+++ dashboard/components/inventory/components/filter/InventoryFilterOperator.tsx
@@ -1,7 +1,7 @@
 import regex from '../../../../utils/regex';
 import Button from '../../../button/Button';
 import Input from '../../../input/Input';
-import { InventoryFilterData } from '../../hooks/useInventory';
+import { InventoryFilterData } from '../../hooks/useInventory/types/useInventoryTypes';
 
 type InventoryFilterOperatorProps = {
   data: InventoryFilterData;
