+++ dashboard/services/settingsService.ts
@@ -167,6 +167,19 @@ const settingsService = {
     }
   },
 
+  async getInventory(urlParams: string, payload: string = '[]') {
+    try {
+      const res = await fetch(
+        `${BASE_URL}/resources/search${urlParams}`,
+        settings('POST', payload)
+      );
+      const data = await res.json();
+      return data;
+    } catch (error) {
+      return Error;
+    }
+  },
+
   async getCustomViewInventory(id: string, urlParams: string, payload: string) {
     try {
       const res = await fetch(
